+++ client/src/containers/AppMain.js
@@ -1,4 +1,5 @@
 import "../index.css";
+import { BrowserRouter as Router, Switch, Route, Link } from "react-router-dom";
 
 import { Marketplace } from "./Marketplace";
 import { Dashboard } from "./Dashboard";
@@ -6,12 +7,23 @@ import { AppHeader } from "./AppHeader";
 
 export const AppMain = () => {
   return (
-    <div>
-      <AppHeader />
-      <div className="app-main">
-        {/* <Marketplace /> */}
-        <Dashboard />
-      </div>
-    </div>
+    <Switch>
+      <Route path="/marketplace">
+        <div>
+          <AppHeader title="Marketplace" />
+          <div className="app-main">
+            <Marketplace />
+          </div>
+        </div>
+      </Route>
+      <Route path="/dashboard">
+        <div>
+          <AppHeader title="Dashboard" />
+          <div className="app-main">
+            <Dashboard />
+          </div>
+        </div>
+      </Route>
+    </Switch>
   );
 };
